requirements.txt
flask
crewai
python-dotenv
gunicorn